fix: Adjust the Pmenu colors
It was really difficult to figure out which popup menu item was selected
versus in the background, especially when there was just two or three
items available. Make the background an accent, and also change the
scrollbar so that it's lightly visible - shouldn't be called out _that_
much, but it's useful to see it.